+++ engine/execution/state/unittest/fixtures.go
@@ -84,7 +84,7 @@ func ComputationResultForBlockFixture(
 		completeBlock.ID(),
 		computationResult.AllChunks(),
 		convertedServiceEvents,
-		[]int{0, 0, 0, 0},
+		[]uint64{0, 0, 0, 0},
 		executionDataID)
 
 	computationResult.ExecutionReceipt = &flow.ExecutionReceipt{
